[FOLD] Prune includes & reduce scope of some functions
