+++ src/ripple/app/ledger/impl/LedgerConsensusImp.cpp
@@ -18,7 +18,6 @@
 //==============================================================================
 
 #include <BeastConfig.h>
-#include <ripple/app/ledger/Consensus.h>
 #include <ripple/app/ledger/InboundLedgers.h>
 #include <ripple/app/ledger/LedgerMaster.h>
 #include <ripple/app/ledger/LedgerTiming.h>
@@ -47,6 +46,158 @@
 #include <type_traits>
 
 namespace ripple {
+
+/** Determines whether the current ledger should close at this time.
+
+    This function should be called when a ledger is open and there is no close
+    in progress, or when a transaction is received and no close is in progress.
+
+    @param anyTransactions indicates whether any transactions have been received
+    @param previousProposers proposers in the last closing
+    @param proposersClosed proposers who have currently closed this ledger
+    @param proposersValidated proposers who have validated the last closed ledger
+    @param previousMSeconds time, in milliseconds, for the previous ledger to
+                            reach consensus (in milliseconds)
+    @param currentMSeconds time, in milliseconds since the previous ledger closed
+    @param openMSeconds time, in milliseconds, since the previous LCL was computed
+    @param idleInterval the network's desired idle interval
+*/
+bool shouldCloseLedger (
+    bool anyTransactions,
+    int previousProposers,
+    int proposersClosed,
+    int proposersValidated,
+    int previousMSeconds,
+    int currentMSeconds,
+    int openMSeconds,
+    int idleInterval)
+{
+    if ((previousMSeconds < -1000) || (previousMSeconds > 600000) ||
+            (currentMSeconds < -1000) || (currentMSeconds > 600000))
+    {
+        WriteLog (lsWARNING, LedgerTiming) <<
+            "shouldCloseLedger Trans=" << (anyTransactions ? "yes" : "no") <<
+            " Prop: " << previousProposers << "/" << proposersClosed <<
+            " Secs: " << currentMSeconds << " (last: " << previousMSeconds << ")";
+        return true;
+    }
+
+    if (!anyTransactions)
+    {
+        // did we miss a transaction?
+        if (proposersClosed > (previousProposers / 4))
+        {
+            WriteLog (lsTRACE, LedgerTiming) <<
+                "no transactions, many proposers: now (" << proposersClosed <<
+                " closed, " << previousProposers << " before)";
+            return true;
+        }
+
+
+        // Only close if we have idled for too long.
+        return currentMSeconds >= (idleInterval * 1000); // normal idle
+    }
+
+    // If we have any transactions, we don't want to close too frequently:
+    if (openMSeconds < LEDGER_MIN_CLOSE)
+    {
+        if ((proposersClosed + proposersValidated) < (previousProposers / 2 ))
+        {
+            WriteLog (lsDEBUG, LedgerTiming) <<
+                "Must wait minimum time before closing";
+            return false;
+        }
+    }
+
+    if (currentMSeconds < previousMSeconds)
+    {
+        if ((proposersClosed + proposersValidated) < previousProposers)
+        {
+            WriteLog (lsDEBUG, LedgerTiming) <<
+                "We are waiting for more closes/validations";
+            return false;
+        }
+    }
+
+    return true;
+}
+
+bool
+checkConsensusReached (int agreeing, int proposing)
+{
+    int currentPercentage = (agreeing * 100) / (proposing + 1);
+
+    return currentPercentage > minimumConsensusPercentage;
+}
+
+/** What state the consensus process is on. */
+enum class ConsensusState
+{
+    No,           // We do not have consensus
+    MovedOn,      // The network has consensus without us
+    Yes           // We have consensus along with the network
+};
+
+/** Determine whether the network reached consensus and whether we joined.
+
+    @param previousProposers proposers in the last closing (not including us)
+    @param currentProposers proposers in this closing so far (not including us)
+    @param currentAgree proposers who agree with us
+    @param currentFinished proposers who have validated a ledger after this one
+    @param previousAgreeTime how long, in milliseconds, it took to agree on the
+                             last ledger
+    @param currentAgreeTime how long, in milliseconds, we've been trying to agree
+*/
+ConsensusState checkConsensus (
+    int previousProposers,
+    int currentProposers,
+    int currentAgree,
+    int currentFinished,
+    int previousAgreeTime,
+    int currentAgreeTime)
+{
+    WriteLog (lsTRACE, LedgerTiming) <<
+        "checkConsensus: prop=" << currentProposers <<
+        "/" << previousProposers <<
+        " agree=" << currentAgree << " validated=" << currentFinished <<
+        " time=" << currentAgreeTime <<  "/" << previousAgreeTime;
+
+    if (currentAgreeTime <= LEDGER_MIN_CONSENSUS)
+        return ConsensusState::No;
+
+    if (currentProposers < (previousProposers * 3 / 4))
+    {
+        // Less than 3/4 of the last ledger's proposers are present; don't
+        // rush: we may need more time.
+        if (currentAgreeTime < (previousAgreeTime + LEDGER_MIN_CONSENSUS))
+        {
+            WriteLog (lsTRACE, LedgerTiming) <<
+                "too fast, not enough proposers";
+            return ConsensusState::No;
+        }
+    }
+
+    // Have we, together with the nodes on our UNL list, reached the treshold
+    // to declare consensus?
+    if (checkConsensusReached (currentAgree + 1, currentProposers))
+    {
+        WriteLog (lsDEBUG, LedgerTiming) << "normal consensus";
+        return ConsensusState::Yes;
+    }
+
+    // Have sufficient nodes on our UNL list moved on and reached the threshold
+    // to declare consensus?
+    if (checkConsensusReached (currentFinished, currentProposers))
+    {
+        WriteLog (lsWARNING, LedgerTiming) <<
+            "We see no consensus, but 80% of nodes have moved on";
+        return ConsensusState::MovedOn;
+    }
+
+    // no consensus yet
+    WriteLog (lsTRACE, LedgerTiming) << "no consensus";
+    return ConsensusState::No;
+}
 
 LedgerConsensusImp::LedgerConsensusImp (
         ConsensusImp& consensus,
